refactor: rename BuilderDrawer to BuilderAndDesignDrawer
