+++ frontend/src/features/admin-form/create/design/components/BuilderAndDesignDrawer/index.ts
@@ -1,3 +1,3 @@
-export * from './BuilderDrawer'
+export * from './BuilderAndDesignDrawer'
 export * from './CreateFieldDrawer'
 export * from './EditFieldDrawer'
